add builder to lotr
