+++ examples/lotr/src/main/scala/com/raphtory/examples/lotr/TutorialRunner.scala
@@ -6,12 +6,12 @@ import com.raphtory.algorithms.generic.NodeList
 import com.raphtory.algorithms.generic.centrality.Degree
 import com.raphtory.algorithms.generic.centrality.PageRank
 import com.raphtory.api.input.Graph.assignID
-import com.raphtory.api.input.ImmutableProperty
-import com.raphtory.api.input.Properties
-import com.raphtory.api.input.Type
+import com.raphtory.api.input.{ImmutableProperty, Properties, Source, Type}
+import com.raphtory.examples.lotr.graphbuilder.LotrGraphBuilder
 import com.raphtory.internals.context.RaphtoryContext
 import com.raphtory.internals.storage.arrow.immutable
 import com.raphtory.sinks.FileSink
+import com.raphtory.spouts.FileSpout
 import com.raphtory.utils.FileUtils
 
 import scala.language.postfixOps
@@ -27,19 +27,22 @@ trait LocalRunner { self: RaphtoryApp =>
       val url  = "https://raw.githubusercontent.com/Raphtory/Data/main/lotr.csv"
       FileUtils.curlFile(path, url)
 
-      val file = scala.io.Source.fromFile(path)
-      file.getLines.foreach { line =>
-        val fileLine   = line.split(",").map(_.trim)
-        val sourceNode = fileLine(0)
-        val srcID      = assignID(sourceNode)
-        val targetNode = fileLine(1)
-        val tarID      = assignID(targetNode)
-        val timeStamp  = fileLine(2).toLong
+      val source = Source(FileSpout(path), LotrGraphBuilder)
+      graph.load(source)
 
-        graph.addVertex(timeStamp, srcID, Properties(ImmutableProperty("name", sourceNode)), Type("Character"))
-        graph.addVertex(timeStamp, tarID, Properties(ImmutableProperty("name", targetNode)), Type("Character"))
-        graph.addEdge(timeStamp, srcID, tarID, Type("Character Co-occurrence"))
-      }
+//      val file = scala.io.Source.fromFile(path)
+//      file.getLines.foreach { line =>
+//        val fileLine   = line.split(",").map(_.trim)
+//        val sourceNode = fileLine(0)
+//        val srcID      = assignID(sourceNode)
+//        val targetNode = fileLine(1)
+//        val tarID      = assignID(targetNode)
+//        val timeStamp  = fileLine(2).toLong
+//
+//        graph.addVertex(timeStamp, srcID, Properties(ImmutableProperty("name", sourceNode)), Type("Character"))
+//        graph.addVertex(timeStamp, tarID, Properties(ImmutableProperty("name", targetNode)), Type("Character"))
+//        graph.addEdge(timeStamp, srcID, tarID, Type("Character Co-occurrence"))
+//      }
 
       //Get simple metrics
       graph
